fault tolerance and resilience in microservices using

+++ movie-rating-service/src/main/java/com/example/movieratingservice/models/Rating.java
@@ -2,9 +2,9 @@ package com.example.movieratingservice.models;
 
 public class Rating {
     private String movieId;
-    private int rating;
+    private Double rating;
 
-    public Rating(String movieId, int rating) {
+    public Rating(String movieId, Double rating) {
         this.movieId = movieId;
         this.rating = rating;
     }
@@ -17,11 +17,11 @@ public class Rating {
         this.movieId = movieId;
     }
 
-    public int getRating() {
+    public Double getRating() {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(Double rating) {
         this.rating = rating;
     }
 }
